+++ src/roswire/proxy/catkin.py
@@ -8,12 +8,14 @@ __all__ = ('CatkinInterface', 'CatkinTools', 'CatkinMake')
 from typing import Optional, List
 import abc
 import shlex
+import os
 
 from loguru import logger
 import attr
 import dockerblade
 
-from ..exceptions import CatkinBuildFailed, CatkinCleanFailed
+from ..exceptions import CatkinBuildFailed, CatkinCleanFailed, \
+    CatkinException
 
 
 class CatkinInterface(abc.ABC):
@@ -22,6 +24,18 @@ class CatkinInterface(abc.ABC):
     @abc.abstractmethod
     def directory(self) -> str:
         """The directory of this catkin workspace."""
+        ...
+
+    @property
+    @abc.abstractmethod
+    def _shell(self) -> dockerblade.shell.Shell:
+        """The shell on the docker container."""
+        ...
+
+    @property
+    @abc.abstractmethod
+    def _files(self) -> dockerblade.files.FileSystem:
+        """The filesystem of the docker container."""
         ...
 
     @abc.abstractmethod
@@ -59,12 +73,34 @@ class CatkinInterface(abc.ABC):
         """
         ...
 
+    def deep_clean(self) -> None:
+        """Removes the build, devel, and install direcotries from the workspace.
+
+        Raises
+        ------
+        CatkinException
+            if removing directories fail.
+        """
+        files = self._files
+        for rm_directory in ['build', 'devel', 'install']:
+            path = os.path.join(self.directory, rm_directory)
+            if files.exists(path):
+                try:
+                    command = f'rm -r {path}'
+                    self._shell.check_output(command, text=True)
+                except dockerblade.exceptions.CalledProcessError as err:
+                    msg = f'Failed to remove directory "{rm_directory}" ' \
+                          f'due to {err}'
+                    logger.error(msg)
+                    raise CatkinException(msg)
+
 
 @attr.s(frozen=True, slots=True, auto_attribs=True)
 class CatkinTools(CatkinInterface):
     """Provides an interface to a catkin workspace created via catkin tools."""
     directory: str
     _shell: dockerblade.shell.Shell
+    _files: dockerblade.files.FileSystem
 
     def clean(self,
               packages: Optional[List[str]] = None,
@@ -84,8 +120,8 @@ class CatkinTools(CatkinInterface):
         logger.debug(f"cleaning via: {command_str}")
         result = shell.run(command_str, cwd=context, text=True)
         duration_mins = result.duration / 60
-        logger.debug("clean completed after %.2f minutes [retcode: %d]:\n%s",
-                     duration_mins, result.returncode, result.output)
+        logger.debug(f"clean completed after {duration_mins:.2f} minutes "
+                     "[retcode: {result.returncode}]:\n{result.output}")
 
         if result.returncode != 0:
             assert isinstance(result.output, str)
@@ -122,8 +158,8 @@ class CatkinTools(CatkinInterface):
                                  time_limit=time_limit,
                                  text=True)
         duration_mins = result.duration / 60
-        logger.debug("build completed after %.2f minutes [retcode: %d]:\n%s",
-                     duration_mins, result.returncode, result.output)
+        logger.debug(f"build completed after {duration_mins:.2f} minutes"
+                     "[retcode: {result.returncode}]:\n{result.output}")
 
         if result.returncode != 0:
             assert isinstance(result.output, str)
@@ -135,13 +171,33 @@ class CatkinMake(CatkinInterface):
     """Provides an interface to a catkin workspace created via catkin_make."""
     directory: str
     _shell: dockerblade.shell.Shell
+    _files: dockerblade.files.FileSystem
 
     def clean(self,
               packages: Optional[List[str]] = None,
               orphans: bool = False,
               context: Optional[str] = None
               ) -> None:
-        raise NotImplementedError
+        shell = self._shell
+        command = ['catkin_make', 'clean']
+        if orphans:
+            raise NotImplementedError
+        if packages:
+            command += ['--pkg']
+            command += [shlex.quote(p) for p in packages]
+        if not context:
+            context = self.directory
+
+        command_str = ' '.join(command)
+        logger.debug(f"cleaning via: {command_str}")
+        result = shell.run(command_str, cwd=context, text=True)
+        duration_mins = result.duration / 60
+        logger.debug(f"clean completed after {duration_mins:.2f} minutes"
+                     "[retcode: {result.returncode}]:\n{result.output}")
+
+        if result.returncode != 0:
+            assert isinstance(result.output, str)
+            raise CatkinCleanFailed(result.returncode, result.output)
 
     def build(self,
               packages: Optional[List[str]] = None,
@@ -153,4 +209,32 @@ class CatkinMake(CatkinInterface):
               context: Optional[str] = None,
               time_limit: Optional[int] = None
               ) -> None:
-        raise NotImplementedError
+        command = ['catkin_make']
+        if packages:
+            command += ['--pkg']
+            command += [shlex.quote(p) for p in packages]
+        if no_deps:
+            raise NotImplementedError
+        if pre_clean:
+            raise NotImplementedError
+        if cmake_args:
+            command += cmake_args
+        if make_args:
+            command += ['--make-args']
+            command += make_args
+        if not context:
+            context = self.directory
+
+        command_str = ' '.join(command)
+        logger.debug(f"building via: {command_str}")
+        result = self._shell.run(command_str,
+                                 cwd=context,
+                                 time_limit=time_limit,
+                                 text=True)
+        duration_mins = result.duration / 60
+        logger.debug(f"build completed after {duration_mins:.2f} minutes "
+                     "[retcode: {result.returncode}]:\n{result.output}")
+
+        if result.returncode != 0:
+            assert isinstance(result.output, str)
+            raise CatkinBuildFailed(result.returncode, result.output)